Fix palette generator region overflow error
The region size the palette generator uses is now clamped to the size
of the scaled image to prevent it from overflowing resulting in
palette generator error.
Reduced the number of symbols in the displayed quote required for
the font size to be decreased to prevent text overflowing the image.