Implement image saving
Added a storage path for storing image files in main.dart.
Wrapped the main display widget in a RepaintBoundary widget.
Added a GlobalKey to the RepaintBoundary.
Added and implemented a method to capture the widget as a png file
in the app directory.
Wrapped the RepaintBoundary in an InkWell for triggering image saving
on long press.