Fix possible error loop in rerolling data sources
When the error occurs and the choosen data source is the last one
from the list, instead of taking the next to last one, takes the first
one in the list.
This will prevent an error loop occuring if the last two sources throw
an error as the default reroll logic is to take the next one in the list
after an error is thrown.